New episodes can now be extracted from the rss feed
We now store a transient hashset of episode guids in the podcast model.
It is built while populating the episode list. We can then check this
set when we want to see if an episode is new or not.

We compare the guids of the list of retrieved episode against the
existing guids. If we have new entries then they must be new episodes.